Add logging of script start and end times in extract_libraryiq.pl
This commit enhances the extract_libraryiq.pl script by adding logging for the script's start and end times. This is achieved by recording the local time when the script begins and when it completes. These timestamps are intended to aid in monitoring and debugging by providing more detailed logging information.

- Added DateTime module usage to track timestamps
- Included log entries for both start and end execution times

Signed-off-by: Ian Skelskey <ianskelskey@gmail.com>